requirements.txt
# Web Framework
flask>=2.3.3
Werkzeug>=2.3.3
Jinja2>=3.1.3

# Database
sqlalchemy>=2.0.23
supabase>=2.0.0

# API & Data Processing
requests>=2.31.0
requests-oauthlib>=1.3.1
python-dotenv>=1.0.0
python-dateutil>=2.8.0
uuid>=1.30

# Data Analysis
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.11.0
plotly>=5.17.0

# Task Scheduling
APScheduler>=3.10.4

# AI/ML
openai>=1.6.0

# Security
Flask-WTF>=1.1.1
gunicorn>=20.1.0

# Development & Testing
pytest>=7.0.0
pytest-flask>=1.2.0
flake8>=6.0.0